Move evaluateGuess into round.js and round-test.js
