+++ test/round-test.js
@@ -3,7 +3,7 @@ const expect = chai.expect;
 
 const { createCard, } = require('../src/card');
 const { createDeck } = require('../src/deck');
-const { createRound, } = require('../src/round');
+const { createRound, evaluateGuess} = require('../src/round');
 
 describe("round", () => {
   let card1, card2, card3, deck, round;
@@ -26,3 +26,33 @@ describe("round", () => {
     expect(round.incorrectGuesses).to.deep.equal([])
   })
 });
+
+describe("guess", () => {
+  it("should be a function", () => {
+    expect(createCard).to.be.a("function");
+  });
+
+  it("should give user feedback if guess is correct", () => {
+    const card = createCard(1, "What allows you to define a set of related information using key-value pairs?", ["object", "array", "function"], "object");
+    const userGuess = "object";
+    let guessResult = evaluateGuess(userGuess, card.correctAnswer);
+    expect(guessResult).to.equal("correct!");
+  });
+
+  it("should give user feedback if guess is incorrect", () => {
+    const card = createCard(1, "What allows you to define a set of related information using key-value pairs?", ["object", "array", "function"], "object");
+    const userGuess = ["array", "function"];
+    let guessResult = evaluateGuess(userGuess, card.correctAnswer);
+    expect(guessResult).to.equal("incorrect");
+  });
+});
+
+describe("turn", () => {
+  
+  it("should be a function", () => {
+    expect(createCard).to.be.a("function");
+  });
+
+  it("should update the turn count")
+
+})
